refactor!: split packages shared between artifacts so modules can resolve
Three packages were each present in two published artifacts. JPMS rejects
that regardless of how the modules are named, so the Automatic-Module-Names
added in the previous commit could not actually be used on the module path:

  ResolutionException: Module org.wiremock.certificate.generator contains
  package com.github.tomakehurst.wiremock.http.ssl, module org.wiremock.core
  exports package com.github.tomakehurst.wiremock.http.ssl to ...

wiremock-core keeps the legacy namespace in every case; the artifact on the
other side of each split moves out:

  certificate-generator
    com.github.tomakehurst.wiremock.http.ssl
      -> com.github.tomakehurst.wiremock.http.ssl.cert
  wiremock (root)
    com.github.tomakehurst.wiremock.standalone
      -> com.github.tomakehurst.wiremock.standalone.cli

The core/junit4 split is resolved from the core side instead. Stubbing and
DslWrapper sat in a .junit package for historical reasons only — Stubbing is
core DSL API implemented by WireMockServer and by junit5's WireMockExtension,
with nothing JUnit-specific about it — so they move to core's .stubbing
package and every existing `import ...wiremock.junit.WireMockRule;` keeps
working.

CertChainAndKey, its fields and constructor, and CertificateAuthority
.generateCertificate were package-private and reached across the artifact
boundary through the split. They are now public, which is unavoidable when
de-splitting in either direction.

Removes wiremock-httpclient-apache5's dependency on certificate-generator:
it only ever used http.ssl types owned by wiremock-core, which the split
package had made impossible for dependency analysis to see.

All thirteen artifacts now resolve together on the module path.
