+++ src/test/java/com/github/tomakehurst/wiremock/StubMappingPersistenceAcceptanceTest.java
@@ -32,8 +32,8 @@ import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.common.FileSource;
 import com.github.tomakehurst.wiremock.common.Json;
 import com.github.tomakehurst.wiremock.common.SingleRootFileSource;
-import com.github.tomakehurst.wiremock.junit.Stubbing;
 import com.github.tomakehurst.wiremock.stubbing.StubMapping;
+import com.github.tomakehurst.wiremock.stubbing.Stubbing;
 import com.github.tomakehurst.wiremock.testsupport.WireMockTestClient;
 import java.io.IOException;
 import java.nio.file.Files;
