+++ src/test/resources/frozen/do-not-throw-generic-exception
@@ -5,4 +5,4 @@ Method <com.github.tomakehurst.wiremock.common.ClasspathFileSource.assertExistsA
 Method <com.github.tomakehurst.wiremock.common.ClasspathFileSource.assertExistsAndIsDirectory()> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (ClasspathFileSource.java:214)
 Method <com.github.tomakehurst.wiremock.jetty.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (JettyHttpServer.java:230)
 Method <com.github.tomakehurst.wiremock.jetty.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (JettyHttpServer.java:220)
-Method <com.github.tomakehurst.wiremock.standalone.CommandLineOptions.captureHelpTextIfRequested(joptsimple.OptionParser)> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (CommandLineOptions.java:241)
+Method <com.github.tomakehurst.wiremock.standalone.cli.CommandLineOptions.captureHelpTextIfRequested(joptsimple.OptionParser)> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (CommandLineOptions.java:241)
