+++ src/test/resources/frozen/unused-methods
@@ -66,8 +66,8 @@ Method <com.github.tomakehurst.wiremock.recording.ProxiedServeEventFilters.setFi
 Method <com.github.tomakehurst.wiremock.recording.ProxiedServeEventFilters.setIds(java.util.List)> is unreferenced in (ProxiedServeEventFilters.java:60)
 Method <com.github.tomakehurst.wiremock.servlet.MultipartRequestConfigElementBuilder.configure(javax.servlet.http.HttpServletRequest)> is unreferenced in (MultipartRequestConfigurer.java:0)
 Method <com.github.tomakehurst.wiremock.servlet.WireMockHttpServletMultipartAdapter.from(javax.servlet.http.Part)> is unreferenced in (WireMockHttpServletMultipartAdapter.java:51)
-Method <com.github.tomakehurst.wiremock.standalone.WireMockServerRunner.port()> is unreferenced in (WireMockServerRunner.java:129)
-Method <com.github.tomakehurst.wiremock.standalone.WireMockServerRunner.stop()> is unreferenced in (WireMockServerRunner.java:115)
+Method <com.github.tomakehurst.wiremock.standalone.cli.WireMockServerRunner.port()> is unreferenced in (WireMockServerRunner.java:129)
+Method <com.github.tomakehurst.wiremock.standalone.cli.WireMockServerRunner.stop()> is unreferenced in (WireMockServerRunner.java:115)
 Method <com.github.tomakehurst.wiremock.stubbing.Scenario.withName(java.lang.String)> is unreferenced in (Scenario.java:144)
 Method <com.github.tomakehurst.wiremock.stubbing.StubImport.stubImport()> is unreferenced in (StubImport.java:44)
 Method <com.github.tomakehurst.wiremock.stubbing.StubMapping.buildJsonStringFor(com.github.tomakehurst.wiremock.stubbing.StubMapping)> is unreferenced in (StubMapping.java:75)
